refactor(telemetry): consume shared anonymized attribution
